Add files via upload
Tackled the issue that the am/pm was not parsing correctly from the input. I found the Regular expression being used to parse the time, and compared it ones i found while researching regular expressions for date and time. I made changes to the Regular expression that now correctly parses the date and time. Depending on how off the formatting of the input is there are still times when it misses the am,pm but works in most cases