Multiple responses using confidence of responses
Adjusted to get the two responses with the highest confidence and combine them into one. Then we check the known possible responses to the response we just created and if the confidence of any of those responses are above a certain level we print the response we created. Otherwise we print a message letting the user know the bot did not know what to say instead of spouting gibberish. The bot needs to be trained for a while and have a good knowledge base of conversations before this logic adapter can be effective.